move startup view to main view
